+++ packages/remix-server-runtime/server.ts
@@ -186,9 +186,7 @@ async function handleDataRequestRR(
         ? error
         : new Error("Unexpected Server Error");
 
-    if (serverMode !== ServerMode.Test && !request.signal.aborted) {
-      console.error(errorInstance);
-    }
+    logServerErrorIfNotAborted(errorInstance, request, serverMode);
 
     return json(serializeError(errorInstance, serverMode), {
       status,
@@ -261,10 +259,7 @@ async function handleDocumentRequestRR(
       requestContext: loadContext,
     });
   } catch (error: unknown) {
-    if (!request.signal.aborted && serverMode !== ServerMode.Test) {
-      console.error(error);
-    }
-
+    logServerErrorIfNotAborted(error, request, serverMode);
     return new Response(null, { status: 500 });
   }
 
@@ -348,6 +343,7 @@ async function handleDocumentRequestRR(
         entryContext
       );
     } catch (error: any) {
+      logServerErrorIfNotAborted(error, request, serverMode);
       return returnLastResortErrorResponse(error, serverMode);
     }
   }
@@ -381,15 +377,22 @@ async function handleResourceRequestRR(
       error.headers.set("X-Remix-Catch", "yes");
       return error;
     }
+    logServerErrorIfNotAborted(error, request, serverMode);
     return returnLastResortErrorResponse(error, serverMode);
   }
 }
 
-function returnLastResortErrorResponse(error: any, serverMode?: ServerMode) {
-  if (serverMode !== ServerMode.Test) {
+function logServerErrorIfNotAborted(
+  error: unknown,
+  request: Request,
+  serverMode: ServerMode
+) {
+  if (serverMode !== ServerMode.Test && !request.signal.aborted) {
     console.error(error);
   }
+}
 
+function returnLastResortErrorResponse(error: any, serverMode?: ServerMode) {
   let message = "Unexpected Server Error";
 
   if (serverMode !== ServerMode.Production) {