+++ packages/remix-css-bundle/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @remix-run/css-bundle
 
+## 1.14.2
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/dev@1.14.2`
+
+## 1.14.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/dev@1.14.1`
+
+## 1.14.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/dev@1.14.0`
+
 ## 1.13.0
 
 ### Patch Changes