+++ packages/remix-netlify/CHANGELOG.md
@@ -1,5 +1,26 @@
 # `@remix-run/netlify`
 
+## 1.14.2
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.14.2`
+
+## 1.14.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.14.1`
+
+## 1.14.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.14.0`
+
 ## 1.13.0
 
 ### Patch Changes